requirements.txt
sendgrid
python-decouple
jinja2
fastapi
uvicorn
pydantic
requests